chore: merge `friends.py` with `blocked_users.py`
